style: Apply style changes based on Google style

+++ src/cat/tests/cat_tests_runner.sh
@@ -50,6 +50,10 @@ function run_test() {
 		echo "Fail"
 		((failed++))
 	fi
+    if ! valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --verbose --tool=memcheck --error-exitcode=1 ./"${BINARY}" $flag $files > /dev/null ; then
+        echo "MEMORY LEAK DETECTED!!!"
+        exit
+    fi
 	rm 1.txt 2.txt
 }
